+++ CHANGELOG.md
@@ -18,3 +18,4 @@ All notable changes to `strapi-wrapper` will be documented in this file.
 - Add ability to set custom endpoint after the collection has been initialised
 - Fix an issue when trying to squash a non array dataset
 - Improve squash behavior for single returned collections
+- Clean up filter return values